fix(stream)!: Rename `input` to `stream`
"Input" feels like the top-level input while "stream" conveys we are
processing as we go.  This ends up clarifying semantics similar to
`Parser::parse` being renamed to `Parser::parse_next`.
