+++ src/_tutorial.rs
@@ -104,7 +104,7 @@
 //! # use winnow::bytes::take_while1;
 //! # use winnow::bytes::tag;
 //! # use winnow::sequence::preceded;
-//! # use winnow::input::AsChar;
+//! # use winnow::stream::AsChar;
 //! struct Request<'s> {
 //!     method: &'s [u8],
 //!     url: &'s [u8],
