+++ src/bytes/streaming.rs
@@ -6,11 +6,11 @@ use crate::error::ErrMode;
 use crate::error::ErrorKind;
 use crate::error::Needed;
 use crate::error::ParseError;
-use crate::input::{
+use crate::lib::std::result::Result::Ok;
+use crate::stream::{
     split_at_offset1_partial, split_at_offset_partial, Compare, CompareResult, ContainsToken,
     FindSlice, Input, Offset, SliceLen, ToUsize,
 };
-use crate::lib::std::result::Result::Ok;
 use crate::{IResult, Parser};
 
 pub(crate) fn any<I, E: ParseError<I>>(input: I) -> IResult<I, <I as Input>::Token, E>
@@ -279,7 +279,7 @@ where
 /// ```rust
 /// # use winnow::{error::ErrMode, error::ErrorKind, error::Needed, IResult};
 /// use winnow::bytes::streaming::take_while;
-/// use winnow::input::AsChar;
+/// use winnow::stream::AsChar;
 ///
 /// fn alpha(s: &[u8]) -> IResult<&[u8], &[u8]> {
 ///   take_while(AsChar::is_alpha)(s)
@@ -331,7 +331,7 @@ where
 /// ```rust
 /// # use winnow::{error::ErrMode, error::{Error, ErrorKind}, error::Needed, IResult};
 /// use winnow::bytes::streaming::take_while1;
-/// use winnow::input::AsChar;
+/// use winnow::stream::AsChar;
 ///
 /// fn alpha(s: &[u8]) -> IResult<&[u8], &[u8]> {
 ///   take_while1(AsChar::is_alpha)(s)
@@ -382,7 +382,7 @@ where
 /// ```rust
 /// # use winnow::{error::ErrMode, error::{Error, ErrorKind}, error::Needed, IResult};
 /// use winnow::bytes::streaming::take_while_m_n;
-/// use winnow::input::AsChar;
+/// use winnow::stream::AsChar;
 ///
 /// fn short_alpha(s: &[u8]) -> IResult<&[u8], &[u8]> {
 ///   take_while_m_n(3, 6, AsChar::is_alpha)(s)
@@ -746,7 +746,7 @@ pub fn escaped<I, Error, F, G, O1, O2>(
 ) -> impl FnMut(I) -> IResult<I, <I as Input>::Slice, Error>
 where
     I: Input + Offset,
-    <I as Input>::Token: crate::input::AsChar,
+    <I as Input>::Token: crate::stream::AsChar,
     F: Parser<I, O1, Error>,
     G: Parser<I, O2, Error>,
     Error: ParseError<I>,
@@ -762,12 +762,12 @@ pub(crate) fn escaped_internal<I, Error, F, G, O1, O2>(
 ) -> IResult<I, <I as Input>::Slice, Error>
 where
     I: Input + Offset,
-    <I as Input>::Token: crate::input::AsChar,
+    <I as Input>::Token: crate::stream::AsChar,
     F: Parser<I, O1, Error>,
     G: Parser<I, O2, Error>,
     Error: ParseError<I>,
 {
-    use crate::input::AsChar;
+    use crate::stream::AsChar;
 
     let mut i = input.clone();
 
@@ -860,8 +860,8 @@ pub fn escaped_transform<I, Error, F, G, Output>(
 ) -> impl FnMut(I) -> IResult<I, Output, Error>
 where
     I: Input + Offset,
-    <I as Input>::Token: crate::input::AsChar,
-    Output: crate::input::Accumulate<<I as Input>::Slice>,
+    <I as Input>::Token: crate::stream::AsChar,
+    Output: crate::stream::Accumulate<<I as Input>::Slice>,
     F: Parser<I, <I as Input>::Slice, Error>,
     G: Parser<I, <I as Input>::Slice, Error>,
     Error: ParseError<I>,
@@ -878,13 +878,13 @@ pub(crate) fn escaped_transform_internal<I, Error, F, G, Output>(
 ) -> IResult<I, Output, Error>
 where
     I: Input + Offset,
-    <I as Input>::Token: crate::input::AsChar,
-    Output: crate::input::Accumulate<<I as Input>::Slice>,
+    <I as Input>::Token: crate::stream::AsChar,
+    Output: crate::stream::Accumulate<<I as Input>::Slice>,
     F: Parser<I, <I as Input>::Slice, Error>,
     G: Parser<I, <I as Input>::Slice, Error>,
     Error: ParseError<I>,
 {
-    use crate::input::AsChar;
+    use crate::stream::AsChar;
 
     let mut offset = 0;
     let mut res = Output::initial(Some(input.input_len()));
@@ -943,7 +943,7 @@ mod tests {
     };
     use crate::error::ErrorKind;
     use crate::error::{ErrMode, Needed};
-    use crate::input::AsChar;
+    use crate::stream::AsChar;
     use crate::IResult;
 
     #[test]
