+++ src/bytes/tests.rs
@@ -8,9 +8,9 @@ use crate::error::ErrMode;
 use crate::error::Error;
 use crate::error::ErrorKind;
 use crate::error::Needed;
-use crate::input::AsChar;
 use crate::multi::length_data;
 use crate::sequence::delimited;
+use crate::stream::AsChar;
 use crate::IResult;
 use crate::Parser;
 use crate::Partial;
