+++ src/lib.rs
@@ -3,7 +3,7 @@
 //! **NOTE:** This is an unofficial, short-lived fork.  I'm nominating winnow to be nom v8.
 //!
 //! nom is a parser combinator library, supporting:
-//! - String (`&str`), byte (`&[u8]`), and [custom input types][crate::input]
+//! - String (`&str`), byte (`&[u8]`), and [custom input types][crate::stream]
 //! - [Partial parsing][crate::Partial]
 //! - Zero copy parsing
 //!
@@ -124,7 +124,7 @@
 //! With functions, you would write it like this:
 //!
 //! ```rust
-//! use winnow::{IResult, bytes::take, input::Partial};
+//! use winnow::{IResult, bytes::take, stream::Partial};
 //! fn take4(input: Partial<&str>) -> IResult<Partial<&str>, &str> {
 //!   take(4u8)(input)
 //! }
@@ -262,7 +262,7 @@
 //!     error::ErrorKind, error::Error, error::Needed,
 //!     number::be_u16,
 //!     bytes::{tag, take},
-//!     input::Partial,
+//!     stream::Partial,
 //! };
 //!
 //! let mut tpl = (be_u16, take(3u8), tag("fg"));
@@ -463,7 +463,7 @@ pub mod error;
 
 mod parser;
 
-pub mod input;
+pub mod stream;
 
 pub mod bits;
 pub mod branch;
@@ -501,7 +501,7 @@ pub mod _tutorial;
 /// }
 /// ```
 pub mod prelude {
-    pub use crate::input::InputIsPartial as _;
+    pub use crate::stream::InputIsPartial as _;
     pub use crate::FinishIResult as _;
     pub use crate::IResult;
     pub use crate::Parser as _;
@@ -509,7 +509,7 @@ pub mod prelude {
 
 pub use error::FinishIResult;
 pub use error::IResult;
-pub use input::Located;
-pub use input::Partial;
-pub use input::Stateful;
 pub use parser::*;
+pub use stream::Located;
+pub use stream::Partial;
+pub use stream::Stateful;
